move all data_ui except tensor to re_data_ui
