+++ crates/re_data_ui/Cargo.toml
@@ -16,7 +16,21 @@ include = ["../../LICENSE-APACHE", "../../LICENSE-MIT", "**/*.rs", "Cargo.toml"]
 all-features = true
 
 [dependencies]
+re_arrow_store.workspace = true
+re_data_store.workspace = true
+re_format.workspace = true
+re_log_types.workspace = true
+re_log.workspace = true
+re_query.workspace = true
+re_renderer.workspace = true
+re_ui.workspace = true
 re_viewer_context.workspace = true
+
+anyhow.workspace = true
+egui_extras.workspace = true
+egui.workspace = true
+itertools.workspace = true
+nohash-hasher.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 puffin.workspace = true
