+++ crates/re_data_ui/src/annotation_context.rs
@@ -1,10 +1,9 @@
 use egui::{color_picker, Vec2};
 use itertools::Itertools;
 use re_log_types::{context::AnnotationInfo, AnnotationContext};
-use re_viewer_context::{UiVerbosity, ViewerContext};
+use re_viewer_context::{auto_color, UiVerbosity, ViewerContext};
 
 use super::DataUi;
-use crate::ui::annotations::auto_color;
 
 const TABLE_SCROLL_AREA_HEIGHT: f32 = 500.0; // add scroll-bars when we get to this height
