+++ crates/re_data_ui/src/component_path.rs
