+++ crates/re_data_ui/src/entity_path.rs
