+++ crates/re_viewer/src/ui/mod.rs
@@ -2,8 +2,7 @@ mod annotations;
 mod auto_layout;
 mod blueprint;
 mod data_blueprint;
-mod item_ui;
-mod scene;
+mod image;
 mod selection_history_ui;
 mod space_view;
 mod space_view_entity_picker;
@@ -15,7 +14,6 @@ mod view_text;
 mod view_time_series;
 mod viewport;
 
-pub(crate) mod data_ui;
 pub(crate) mod memory_panel;
 pub(crate) mod selection_panel;
 pub(crate) mod time_panel;
@@ -24,10 +22,8 @@ pub mod view_spatial;
 
 // ----
 
-use self::scene::SceneQuery;
-
 pub(crate) use self::blueprint::Blueprint;
-pub(crate) use self::space_view::SpaceView;
+pub(crate) use self::space_view::{item_ui, SpaceView};
 
 pub use self::annotations::{Annotations, DefaultColor, MISSING_ANNOTATIONS};
 pub use self::view_category::ViewCategory;
