+++ crates/re_viewer/src/ui/time_panel/mod.rs
@@ -9,12 +9,13 @@ use std::ops::RangeInclusive;
 use egui::{pos2, Color32, CursorIcon, NumExt, PointerButton, Rect, Shape, Vec2};
 
 use re_data_store::{EntityTree, InstancePath, TimeHistogram};
+use re_data_ui::{item_ui, DataUi};
 use re_log_types::{ComponentPath, EntityPathPart, TimeInt, TimeRange, TimeReal};
 use re_viewer_context::{Item, TimeControl, TimeView, UiVerbosity, ViewerContext};
 
 use crate::misc::TimeControlUi;
 
-use super::{data_ui::DataUi, item_ui, selection_panel::what_is_selected_ui, Blueprint};
+use super::{selection_panel::what_is_selected_ui, Blueprint};
 
 use time_axis::TimelineAxis;
 use time_ranges_ui::TimeRangesUi;
