+++ crates/re_viewer/src/ui/view_bar_chart/scene.rs
@@ -5,9 +5,7 @@ use re_data_store::EntityPath;
 use re_log::warn_once;
 use re_log_types::component_types::{self, InstanceKey, Tensor};
 use re_query::query_entity_with_primary;
-use re_viewer_context::ViewerContext;
-
-use crate::ui::scene::SceneQuery;
+use re_viewer_context::{SceneQuery, ViewerContext};
 
 /// A bar chart scene, with everything needed to render it.
 #[derive(Default)]
