+++ crates/re_viewer/src/ui/view_spatial/scene/scene_part/arrows3d.rs
@@ -5,11 +5,11 @@ use re_log_types::{
 };
 use re_query::{query_primary_with_history, EntityView, QueryError};
 use re_renderer::{renderer::LineStripFlags, Size};
-use re_viewer_context::ViewerContext;
+use re_viewer_context::{SceneQuery, ViewerContext};
 
 use crate::{
     misc::{SpaceViewHighlights, TransformCache},
-    ui::{scene::SceneQuery, view_spatial::SceneSpatial, DefaultColor},
+    ui::{view_spatial::SceneSpatial, DefaultColor},
 };
 
 use super::{instance_key_to_picking_id, ScenePart};
