+++ crates/re_viewer/src/ui/view_spatial/scene/scene_part/meshes.rs
@@ -4,12 +4,11 @@ use re_log_types::{
     Component, Mesh3D,
 };
 use re_query::{query_primary_with_history, EntityView, QueryError};
-use re_viewer_context::ViewerContext;
+use re_viewer_context::{SceneQuery, ViewerContext};
 
 use crate::{
     misc::{caches::MeshCache, SpaceViewHighlights, TransformCache},
     ui::{
-        scene::SceneQuery,
         view_spatial::{MeshSource, SceneSpatial},
         DefaultColor,
     },
