+++ crates/re_viewer/src/ui/view_text/mod.rs
@@ -2,4 +2,4 @@ mod scene;
 pub(crate) use self::scene::{SceneText, TextEntry};
 
 mod ui;
-pub(crate) use self::ui::{level_to_rich_text, view_text, ViewTextState};
+pub(crate) use self::ui::{view_text, ViewTextState};
