+++ crates/re_viewer/src/ui/view_text/scene.rs
@@ -5,9 +5,7 @@ use re_log_types::{
     Component, RowId,
 };
 use re_query::{range_entity_with_primary, QueryError};
-use re_viewer_context::ViewerContext;
-
-use crate::ui::SceneQuery;
+use re_viewer_context::{SceneQuery, ViewerContext};
 
 use super::ui::ViewTextFilters;
