+++ crates/re_viewer_context/src/lib.rs
@@ -6,20 +6,24 @@ mod app_options;
 mod caches;
 mod component_ui_registry;
 mod item;
+mod scene_query;
 mod selection_history;
 mod selection_state;
 mod time_control;
+mod utils;
 mod viewer_context;
 
 pub use app_options::AppOptions;
 pub use caches::{Cache, Caches};
 pub use component_ui_registry::{ComponentUiRegistry, UiVerbosity};
 pub use item::{Item, ItemCollection};
+pub use scene_query::SceneQuery;
 pub use selection_history::SelectionHistory;
 pub use selection_state::{
     HoverHighlight, HoveredSpace, InteractionHighlight, SelectionHighlight, SelectionState,
 };
 pub use time_control::{Looping, PlayState, TimeControl, TimeView};
+pub use utils::{auto_color, level_to_rich_text};
 pub use viewer_context::{RecordingConfig, ViewerContext};
 
 // ---------------------------------------------------------------------------
