+++ crates/re_viewer_context/src/scene_query.rs
@@ -2,8 +2,6 @@ use nohash_hasher::IntSet;
 
 use re_data_store::{EntityPath, EntityProperties, EntityPropertyMap, TimeInt, Timeline};
 
-// ---
-
 pub struct SceneQuery<'s> {
     pub entity_paths: &'s IntSet<EntityPath>,
     pub timeline: Timeline,
@@ -15,7 +13,7 @@ impl<'s> SceneQuery<'s> {
     /// Iter over all of the currently visible [`EntityPath`]s in the [`SceneQuery`].
     ///
     /// Also includes the corresponding [`EntityProperties`].
-    pub(crate) fn iter_entities(&self) -> impl Iterator<Item = (&EntityPath, EntityProperties)> {
+    pub fn iter_entities(&self) -> impl Iterator<Item = (&EntityPath, EntityProperties)> {
         self.entity_paths
             .iter()
             .map(|entity_path| (entity_path, self.entity_props_map.get(entity_path)))
