+++ crates/re_viewer_context/src/utils/mod.rs
@@ -1,0 +1,5 @@
+mod color;
+mod text;
+
+pub use color::auto_color;
+pub use text::level_to_rich_text;
