+++ crates/re_viewer_context/src/utils/text.rs
@@ -1,0 +1,15 @@
+use egui::{Color32, RichText};
+
+pub fn level_to_rich_text(ui: &egui::Ui, lvl: &str) -> RichText {
+    match lvl {
+        "CRITICAL" => RichText::new(lvl)
+            .color(Color32::WHITE)
+            .background_color(ui.visuals().error_fg_color),
+        "ERROR" => RichText::new(lvl).color(ui.visuals().error_fg_color),
+        "WARN" => RichText::new(lvl).color(ui.visuals().warn_fg_color),
+        "INFO" => RichText::new(lvl).color(Color32::LIGHT_GREEN),
+        "DEBUG" => RichText::new(lvl).color(Color32::LIGHT_BLUE),
+        "TRACE" => RichText::new(lvl).color(Color32::LIGHT_GRAY),
+        _ => RichText::new(lvl).color(ui.visuals().text_color()),
+    }
+}
